delete unused NoYields
