--- lib/mocha/no_yields.rb
@@ -1,5 +1,0 @@
-module Mocha
-  class NoYields
-    def each; end
-  end
-end
