+++ lib/mocha/yield_parameters.rb
@@ -1,4 +1,3 @@
-require 'mocha/no_yields'
 require 'mocha/multiple_yields'
 
 module Mocha
