+++ test/unit/yield_parameters_test.rb
@@ -1,7 +1,6 @@
 require File.expand_path('../../test_helper', __FILE__)
 
 require 'mocha/yield_parameters'
-require 'mocha/no_yields'
 require 'mocha/multiple_yields'
 
 class YieldParametersTest < Mocha::TestCase
